+++ modules/gpu/include/correspondcuda.h
@@ -75,8 +75,8 @@ private:
 	void findEpipolarMatchesManyThreads(lpt::ImageFrameGroup& frame_group); //TODO: try to make this work!!
 	int getNextComputeDeviceID();
 	thrust::host_vector<CameraPairCUDA> camera_pairs_h;
-	thrust::host_vector<int> num_matches_h;
-	thrust::device_vector<int> num_matches_d;
+	thrust::host_vector<size_t> num_matches_h;
+	thrust::device_vector<size_t> num_matches_d;
 	thrust::host_vector< MatchIDs > matches2way_h;
 	thrust::device_vector< MatchIDs > matches2way_d;
 	
@@ -90,8 +90,8 @@ private:
 	thrust::device_vector<float> particles_x_d;
 	thrust::device_vector<float> particles_y_d;
 
-	thrust::host_vector<int> num_particles_h;
-	thrust::device_vector<int> num_particles_d;
+	thrust::host_vector<size_t> num_particles_h;
+	thrust::device_vector<size_t> num_particles_d;
 
 	boost::mutex mutex;
 	queue<int> compute_devices_available;
